suite-sparse: fix CHOLMOD in 4.1.0 with -DNCAMD, change dep to metis4
In SuiteSparse 4.1.0, the -DNPARTITION option was partially split out in
to -DNCAMD. However, looks like there's a bug in the Makefile, so when METIS
is not linked, libcholmod.a does not get properly built due to an error in
CAMD-related build setup. Add -DNCAMD to match the 4.0.2 behavior and get
it building again.

Also change the metis dependency to use the metis4 formula instead of metis,
since SuiteSparse is not compatible with Metix 5.x.